Closed unnecessary AuthenticationService class
